seperate simply render_template from blueprint, just use app and build a views package to store them
