+++ taipei-day-trip/TaipeiTravel/attraction.py
@@ -9,17 +9,6 @@ attraction_bp = Blueprint("attraction_bp",
 						  static_folder="static")
 
 
-# attraction pages
-@attraction_bp.route("/")
-def index():
-	return render_template("index.html")
-
-
-@attraction_bp.route("/attraction/<id>")
-def attraction(id):
-	return render_template("attraction.html")
-
-
 # ------ Attraction API -------
 # build funfciton for json format 
 def to_dict(attraction_result:list, image_result:dict):
