--- taipei-day-trip/TaipeiTravel/thankyou.py
@@ -1,8 +1,0 @@
-from flask import render_template, Blueprint
-
-thankyou_bp = Blueprint("thankyou_bp", __name__)
-
-# Pages
-@thankyou_bp.route("/thankyou")
-def thankyou():
-	return render_template("thankyou.html")
