+++ taipei-day-trip/TaipeiTravel/views/static/attraction.js
