+++ taipei-day-trip/TaipeiTravel/views/static/booking.js
