+++ taipei-day-trip/TaipeiTravel/views/static/image/mrtbar-leftArrow-hover.png
