+++ taipei-day-trip/TaipeiTravel/views/static/index.js
