+++ taipei-day-trip/TaipeiTravel/views/static/thankyou.js
